Address PR feedback on graph controls deps and prefigure readiness
