+++ packages/doenetml/src/Viewer/renderers/hooks/useGraphControlsState.tsx
@@ -6,13 +6,15 @@ import {
     GraphControlsMode,
     deriveActiveInputKeys,
     deriveActiveSliderAxisKeys,
-    formatCoordinateForControls,
     makeInputKey,
     normalizedSliderBounds,
-    parseOrderedPair,
-    parseSingleMathNumber,
     pruneRecordByActiveKeys,
 } from "../utils/graphControls";
+import {
+    formatCoordinateForControls,
+    parseOrderedPair,
+    parseSingleMathNumber,
+} from "../utils/graphControlsMath";
 
 type UseGraphControlsStateArgs = {
     id: string;
